refining & moving search bar
