+++ project/components/GlobalNav.vue
@@ -10,7 +10,6 @@
 
             <v-spacer></v-spacer>
 
-            <v-text-field v-if="firebaseUser" hide-details prepend-icon="mdi-magnify" single-line></v-text-field>
 
             <v-btn v-if="firebaseUser" to="/review" class="text-none text-subtitle-1" color="#5865f2" size="small"
                 variant="flat">
@@ -37,6 +36,9 @@
 </template>
 
 <script setup>
+//for search bar in nav
+import { queryEntireCollection } from '~/lib/db';
+
 const firebaseUser = useFirebaseUser();
 
 const tabItems = [
@@ -57,4 +59,27 @@ function onTabClick(tab) {
         navigateTo('/profile');
     }
 }
+
+
+// const allCourses = ref();
+
+// const selected = ref();
+
+// onMounted(async () => {
+//     allCourses.value = await queryEntireCollection('classes');
+// }
+// );
+
+
+// // watch works directly on a ref
+// watch(selected, async () => {
+//     await navigateTo({
+//         path: '/details/',
+//         query: {
+//             id: selected.value.id,
+//         },
+//         replace: true,
+//     })
+// })
+
 </script>
